remove stale bs92 erosion code; stitch modified code
